Move TargetTracker outside of Transposer
