+++ com.unity.cinemachine/Editor/Editors/CinemachineFreeLookEditor.cs
@@ -23,7 +23,7 @@ namespace Cinemachine
             excluded.Add(FieldPath(x => x.m_Orbits));
             if (!Target.m_CommonLens)
                 excluded.Add(FieldPath(x => x.m_Lens));
-            if (Target.m_BindingMode == CinemachineTransposer.BindingMode.SimpleFollowWithWorldUp)
+            if (Target.m_BindingMode == TargetTracker.BindingMode.SimpleFollowWithWorldUp)
             {
                 excluded.Add(FieldPath(x => x.m_Heading));
                 excluded.Add(FieldPath(x => x.m_RecenterToTargetHeading));
@@ -67,7 +67,7 @@ namespace Cinemachine
             }
             
             Target.m_XAxis.ValueRangeLocked
-                = (Target.m_BindingMode == CinemachineTransposer.BindingMode.SimpleFollowWithWorldUp);
+                = (Target.m_BindingMode == TargetTracker.BindingMode.SimpleFollowWithWorldUp);
 
             // Ordinary properties
             BeginInspector();
