+++ com.unity.cinemachine/Editor/Editors/CinemachineTransposerEditor.cs
@@ -17,8 +17,8 @@ namespace Cinemachine.Editor
             switch (Target.m_BindingMode)
             {
                 default:
-                case CinemachineTransposer.BindingMode.LockToTarget:
-                    if (Target.m_AngularDampingMode == CinemachineTransposer.AngularDampingMode.Euler)
+                case TargetTracker.BindingMode.LockToTarget:
+                    if (Target.m_AngularDampingMode == TargetTracker.AngularDampingMode.Euler)
                         excluded.Add(FieldPath(x => x.m_AngularDamping));
                     else
                     {
@@ -27,26 +27,26 @@ namespace Cinemachine.Editor
                         excluded.Add(FieldPath(x => x.m_RollDamping));
                     }
                     break;
-                case CinemachineTransposer.BindingMode.LockToTargetNoRoll:
+                case TargetTracker.BindingMode.LockToTargetNoRoll:
                     excluded.Add(FieldPath(x => x.m_RollDamping));
                     excluded.Add(FieldPath(x => x.m_AngularDamping));
                     excluded.Add(FieldPath(x => x.m_AngularDampingMode));
                     break;
-                case CinemachineTransposer.BindingMode.LockToTargetWithWorldUp:
+                case TargetTracker.BindingMode.LockToTargetWithWorldUp:
                     excluded.Add(FieldPath(x => x.m_PitchDamping));
                     excluded.Add(FieldPath(x => x.m_RollDamping));
                     excluded.Add(FieldPath(x => x.m_AngularDamping));
                     excluded.Add(FieldPath(x => x.m_AngularDampingMode));
                     break;
-                case CinemachineTransposer.BindingMode.LockToTargetOnAssign:
-                case CinemachineTransposer.BindingMode.WorldSpace:
+                case TargetTracker.BindingMode.LockToTargetOnAssign:
+                case TargetTracker.BindingMode.WorldSpace:
                     excluded.Add(FieldPath(x => x.m_PitchDamping));
                     excluded.Add(FieldPath(x => x.m_YawDamping));
                     excluded.Add(FieldPath(x => x.m_RollDamping));
                     excluded.Add(FieldPath(x => x.m_AngularDamping));
                     excluded.Add(FieldPath(x => x.m_AngularDampingMode));
                     break;
-                case CinemachineTransposer.BindingMode.SimpleFollowWithWorldUp:
+                case TargetTracker.BindingMode.SimpleFollowWithWorldUp:
                     excluded.Add(FieldPath(x => x.m_XDamping));
                     excluded.Add(FieldPath(x => x.m_PitchDamping));
                     excluded.Add(FieldPath(x => x.m_YawDamping));
