+++ com.unity.cinemachine/Runtime/Core/TargetTracker.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0aa26b55dba50a24ba30bb759b2db4e9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
